+++ plugins/content-manager/src/test/java/com/wazuh/contentmanager/cti/catalog/service/UpdateServiceImplTests.java
@@ -20,11 +20,11 @@ import com.fasterxml.jackson.databind.JsonNode;
 
 import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
 import org.apache.hc.core5.http.ContentType;
-import org.junit.Assert;
 import org.opensearch.action.get.GetResponse;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.test.OpenSearchTestCase;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 
 import java.nio.charset.StandardCharsets;